Monitoring: Move `TablePagination` to its own file
